integrate first draft of menu
split the game state into menu and level

+++ Tabulo/Control/fgTabuloEvent.h
@@ -15,6 +15,6 @@
 
 @property (readonly) NSUInteger Level;
 
-- (id)init:(NSUInteger)_event level:(NSUInteger)_index;
+- (id)init:(unsigned int)_event level:(NSUInteger)_index;
 
 @end
